Update ART integration test suite
	- Refacto and split art test integration suite into mutliple files
	- GNSS serial test:
		- Add new required messages for GNSS
		- Configure gnss receiver during integration test suite
		- Check for valid 3D fix to validate gnss receiver
		- Do not print Gnss messages received if already received
	- PTP device test:
		- Update Threshold for adjtime
		- Delete some adjtime tests
	- Phase error tracking test:
		- Launch oscillatord service from ART Test integration suite
		- Add libsystemd to dependencies
		- Test checks phase error for a period a time
		- If absolute phase error is above the limit, calibration
		requested before testing again
		- If phase error is still above limit test fails
	- Update extts test program
	- Add prefix to Monitoring thread logs
	- Update CMake files for tests

+++ src/monitoring.h
@@ -18,7 +18,7 @@ struct monitoring {
 	struct oscillator_ctrl ctrl_values;
 	const char *oscillator_model;
 	int disciplining_status;
-	int phase_error;
+	int64_t phase_error;
 	int fix;
 	int satellites_count;
 	int leap_seconds;
